OCSADV-255 Move all implicit model conversions to ModelConverter #time 10m
