+++ bundle/edu.gemini.ags/src/main/scala/edu/gemini/ags/impl/package.scala
@@ -5,48 +5,18 @@ import edu.gemini.catalog.api.MagnitudeLimits.{FaintnessLimit, SaturationLimit}
 import edu.gemini.shared.util.immutable.PredicateOp
 import edu.gemini.shared.util.immutable.ScalaConverters.ScalaOptionOps
 import edu.gemini.pot.ModelConverters._
-import edu.gemini.spModel.core._
 import edu.gemini.spModel.obs.context.ObsContext
 import edu.gemini.spModel.rich.shared.immutable._
 import edu.gemini.spModel.target.SPTarget
 import edu.gemini.spModel.target.env.GuideProbeTargets
 
 import edu.gemini.skycalc
-import edu.gemini.shared.skyobject
 import edu.gemini.spModel.target.system.NonSiderealTarget
 
 import scalaz._
 import Scalaz._
 
 package object impl {
-
-  implicit class OldAngle2New(val angle: skycalc.Angle) extends AnyVal{
-    def toNewModel: Angle = Angle.fromDegrees(angle.toDegrees.getMagnitude)
-  }
-
-  implicit class NewAngle2Old(val angle: Angle) extends AnyVal {
-    def toOldModel: skycalc.Angle = skycalc.Angle.degrees(angle.toDegrees)
-  }
-
-  implicit class OldOffset2New(val offset: skycalc.Offset) extends AnyVal {
-    def toNewModel: Offset = Offset(offset.p().toNewModel, offset.q().toNewModel)
-  }
-
-  implicit class OldCoordinates2New(val c: skycalc.Coordinates) extends AnyVal {
-    def toNewModel: Coordinates = Coordinates(RightAscension.fromAngle(c.getRa.toNewModel), Declination.fromAngle(c.getDec.toNewModel).getOrElse(Declination.zero))
-  }
-
-  implicit class OldMagnitude2New(val m: skyobject.Magnitude) extends AnyVal {
-    def toNewModel: Magnitude = new Magnitude(m.getBrightness, m.getBand.toNewModel)
-  }
-
-  implicit class NewMagnitude2Old(val m: Magnitude) extends AnyVal {
-    def toOldModel: skyobject.Magnitude = new skyobject.Magnitude(m.band.toOldModel, m.value)
-  }
-
-  implicit class HmsDegCoords2Coordinates(val c: skyobject.coords.HmsDegCoordinates) extends AnyVal {
-    def toNewModel: Coordinates = Coordinates(RightAscension.fromAngle(c.getRa.toNewModel), Declination.fromAngle(c.getDec.toNewModel).getOrElse(Declination.zero))
-  }
 
   // REMOVE When AGS is fully ported
   @Deprecated
